chore: move user controller and service to a separate folder
